Update FEH_summoning_header_ver5.R
Fixed bug related to the generation of 4-star focus heroes. When a generating a circle for a single 4-star focus hero that isn't in the first position of the array (for instance, position 2), the script would assume more 4-star focus heroes that occupy all the positions up to that position (so positions 1 and 2). Script should now correctly generate 4-star focus heroes.