Add files via upload
Created a file with a lot of info about each hero, for those who are interested. Skill information (including weapons) are not yet included for new bridal banner.

I plan to keep this file updated, but it will likely always be missing some info for the newest heroes due to how I create it.